refactor(domain): move OrchestrationResult into voom-domain with serde derives (#378)
