+++ crates/voom-domain/src/lib.rs
@@ -6,6 +6,7 @@ pub mod capability_map;
 pub mod capability_resolution;
 pub mod compiled;
 pub mod evaluation;
+pub mod orchestration;
 pub mod errors;
 pub mod estimate;
 pub mod events;
